# scaffolds/blueprints/api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/accounts/{id}": {
      "get": {
        "description": "Fetch detailed account information including the 5 most recent users, entities, and roles with counts.",
        "operationId": "AccountController_fetchAccount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Account ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detailed account information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchAccountDeepResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Failed to get account details"
          },
          "404": {
            "description": "Account not found"
          }
        },
        "summary": "Fetch account details",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/accounts/{id}/users": {
      "get": {
        "description": "Fetch users linked to an account with pagination and filtering options.",
        "operationId": "AccountController_fetchAccountUsers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Account ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term to filter users",
            "schema": {
              "example": "john",
              "type": "string"
            }
          },
          {
            "name": "roleIds",
            "required": false,
            "in": "query",
            "description": "Filter by role IDs",
            "schema": {
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          },
          {
            "name": "entityIds",
            "required": false,
            "in": "query",
            "description": "Filter by entity IDs",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "Order by field",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": [
                "createdAt",
                "updatedAt",
                "name",
                "lastname"
              ]
            }
          },
          {
            "name": "includeDirectUsers",
            "required": false,
            "in": "query",
            "description": "Include users linked directly to the account",
            "schema": {
              "default": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchAccountUsersResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Failed to fetch account users"
          },
          "404": {
            "description": "Account not found"
          }
        },
        "summary": "Fetch account users",
        "tags": [
          "Accounts"
        ]
      },
      "patch": {
        "description": "Update the users linked to an account.",
        "operationId": "AccountController_updateAccountUsers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Account ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountUsersDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The users have been successfully updated for the account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateAccountUsersResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Failed to update account users"
          },
          "401": {
            "description": "User does not have permission to manage the account"
          },
          "404": {
            "description": "Account not found"
          }
        },
        "summary": "Update account users",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/accounts/{id}/entities": {
      "get": {
        "description": "Fetch entities linked to an account with pagination and filtering options.",
        "operationId": "AccountController_fetchAccountEntities",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Account ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term to filter entities",
            "schema": {
              "example": "company",
              "type": "string"
            }
          },
          {
            "name": "userIds",
            "required": false,
            "in": "query",
            "description": "Filter by user IDs (entities with these users)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "includeInactiveUsers",
            "required": false,
            "in": "query",
            "description": "Include users that are not active",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "Order by field",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": [
                "createdAt",
                "updatedAt",
                "name",
                "organizationName"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of entities",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchAccountEntitiesResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Failed to fetch account entities"
          },
          "404": {
            "description": "Account not found"
          }
        },
        "summary": "Fetch account entities",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/accounts/{id}/roles": {
      "get": {
        "description": "Fetch roles linked to an account with pagination and filtering options.",
        "operationId": "AccountController_fetchAccountRoles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Account ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term to filter roles",
            "schema": {
              "example": "admin",
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "Order by field",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": [
                "createdAt",
                "name"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of roles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchAccountRolesResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Failed to fetch account roles"
          },
          "404": {
            "description": "Account not found"
          }
        },
        "summary": "Fetch account roles",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/accounts/{id}/status": {
      "patch": {
        "description": "Activate or deactivate an account that the user has access to.",
        "operationId": "AccountController_updateAccountStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Account ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateAccountStatusResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Failed to update account status"
          },
          "401": {
            "description": "User does not have permission to manage the account"
          },
          "404": {
            "description": "Account not found"
          }
        },
        "summary": "Update account status",
        "tags": [
          "Accounts"
        ]
      }
    },
...
  },
  "info": {
    "title": "SaaSFoundry API",
    "description": "An open-source solution for managing clients, invoices, and financial tasks.",
    "version": "1.0.0",
    "contact": {
      "name": "SaaSFoundry Team",
      "url": "https://github.com/DiamondForgeFr/SaaSFoundry",
      "email": "anthony.gachet@diamondforge.fr"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PeopleDto": {
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string",
            "nullable": true,
            "description": "First name",
            "example": "John"
          },
          "lastname": {
            "type": "string",
            "nullable": true,
            "description": "Last name",
            "example": "Doe"
          }
        },
        "required": [
          "firstname",
          "lastname"
        ]
      },
...
      "OrganizationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Organization unique identifier",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Organization name",
            "example": "ACME Corporation"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
      "CollectionResponseDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number",
            "description": "Total number of items in the collection"
          },
          "values": {
            "description": "Collection items",
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "count",
          "values"
        ]
      },
      "FetchAccountDeepResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Account ID",
            "example": "cmagp5dy70001t84nzb9t39j6"
          },
          "name": {
            "type": "string",
            "description": "Account name",
            "example": "Acme Corp"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Account description",
            "example": "Main account for Acme Corporation"
          },
          "isActive": {
            "type": "boolean",
            "description": "Account active status",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation date",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update date",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "users": {
            "description": "Account users",
            "allOf": [
              {
                "$ref": "#/components/schemas/CollectionResponseDto"
              }
            ]
          },
          "entities": {
            "description": "Account entities",
            "allOf": [
              {
                "$ref": "#/components/schemas/CollectionResponseDto"
              }
            ]
          },
          "roles": {
            "description": "Account roles",
            "allOf": [
              {
                "$ref": "#/components/schemas/CollectionResponseDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "isActive",
          "createdAt",
          "updatedAt",
          "users",
          "entities",
          "roles"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "pagination": {
            "type": "object",
            "properties": {
              "current": {
                "type": "number"
              },
              "limit": {
                "type": "number"
              },
              "total": {
                "type": "number"
              }
            },
            "required": [
              "current",
              "limit",
              "total"
            ]
          },
          "count": {
            "type": "number",
            "description": "Total number of items",
            "example": 42
          }
        },
        "required": [
          "pagination",
          "count"
        ]
      },
      "FetchAccountUsersResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Users data",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountUserDto"
            }
          },
          "meta": {
            "description": "Pagination metadata",
            "allOf": [
              {
                "$ref": "#/components/schemas/PaginationMetaDto"
              }
            ]
          }
        },
        "required": [
          "items",
          "meta"
        ]
      },
      "FetchAccountEntitiesResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "List of entities",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityWithOrganizationDto"
            }
          },
          "meta": {
            "description": "Pagination metadata",
            "allOf": [
              {
                "$ref": "#/components/schemas/PaginationMetaDto"
              }
            ]
          }
        },
        "required": [
          "items",
          "meta"
        ]
      },
      "FetchAccountRolesResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "List of roles",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountRoleDto"
            }
          },
          "meta": {
            "description": "Pagination metadata",
            "allOf": [
              {
                "$ref": "#/components/schemas/PaginationMetaDto"
              }
            ]
          }
        },
        "required": [
          "items",
          "meta"
        ]
      },
      "UpdateAccountStatusDto": {
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean",
            "description": "New account status",
            "example": true
          }
        },
        "required": [
          "isActive"
        ]
      },
      "UpdateAccountStatusResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Account ID"
          },
          "name": {
            "type": "string",
            "description": "Account name"
          },
          "isActive": {
            "type": "boolean",
            "description": "Account active status"
          }
        },
        "required": [
          "id",
          "name",
          "isActive"
        ]
      },
      "UpdateAccountUsersDto": {
        "type": "object",
        "properties": {
          "userIds": {
            "description": "List of user IDs to associate with the account",
            "example": [
              "cmagp5dy70001t84nzb9t39j6",
              "cmagp5dy70001t84nzb9t39j7"
            ],
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "/^c[a-z0-9]{20,}$/"
            }
          }
        },
        "required": [
          "userIds"
        ]
      },
      "AccountUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          },
          "email": {
            "type": "string",
            "description": "User email",
            "example": "john.doe@example.com"
          },
          "isActive": {
            "type": "boolean",
            "description": "User active status",
            "example": true
          },
          "people": {
            "nullable": true,
            "description": "User personal information",
            "allOf": [
              {
                "$ref": "#/components/schemas/PeopleDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "email",
          "isActive",
          "people"
        ]
      },
      "UpdateAccountUsersResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The ID of the account",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          },
          "name": {
            "type": "string",
            "description": "Account name"
          },
          "users": {
            "description": "List of users linked to the account",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountUserDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "users"
        ]
      },
...
      "EntityWithOrganizationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Entity ID",
            "example": "cmagp5dy70001t84nzb9t39j8"
          },
          "name": {
            "type": "string",
            "description": "Entity name",
            "example": "Main Office"
          },
          "organization": {
            "description": "Associated organization",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrganizationDto"
              }
            ],
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "organization"
        ]
      },
      "AccountRoleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Role ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Role name",
            "example": "Admin"
          },
          "description": {
            "type": "string",
            "description": "Role description",
            "example": "Administrator with full access",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "description": "Role active status",
            "example": true
          },
          "isGlobal": {
            "type": "boolean",
            "description": "Whether the role is global",
            "example": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation date",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update date",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "isActive",
          "isGlobal",
          "createdAt",
          "updatedAt"
        ]
      }
    }
  }
}