# scaffolds/overlays/monorepo/root/packages/api-client/src/generated/api/model/accountUserDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SaaSFoundry API
 * An open-source solution for managing clients, invoices, and financial tasks.
 * OpenAPI spec version: 1.0.0
 */
import type { AccountUserDtoPeople } from './accountUserDtoPeople';

export interface AccountUserDto {
  /** User ID */
  id: string;
  /** User email */
  email: string;
  /** User active status */
  isActive: boolean;
  /**
   * User personal information
   * @nullable
   */
  people: AccountUserDtoPeople;
}