# scaffolds/overlays/monorepo/root/packages/api-client/src/generated/api/model/fetchAccountRolesResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SaaSFoundry API
 * An open-source solution for managing clients, invoices, and financial tasks.
 * OpenAPI spec version: 1.0.0
 */
import type { AccountRoleDto } from './accountRoleDto';
import type { PaginationMetaDto } from './paginationMetaDto';

export interface FetchAccountRolesResponseDto {
  /** List of roles */
  items: AccountRoleDto[];
  /** Pagination metadata */
  meta: PaginationMetaDto;
}