# scaffolds/overlays/monorepo/root/packages/api-client/src/generated/api/model/organizationDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SaaSFoundry API
 * An open-source solution for managing clients, invoices, and financial tasks.
 * OpenAPI spec version: 1.0.0
 */

export interface OrganizationDto {
  /** Organization unique identifier */
  id: string;
  /** Organization name */
  name: string;
}